Internals

Keccak

KangarooTwelve.keccak_p1600 — Function

keccak_p1600(state::NTuple{25, UInt64}, ::Val{nrounds}=Val{12}())

Apply the Keccak-p[nrounds, 1600] permutation to state. This is formally defined in the Keccak reference and formalised in FIPS 202.

Turboshake

KangarooTwelve.turboshake — Function

turboshake(output::Type, message::AbstractVector{<:UInt8to64},
           delimsuffix::UInt8=0x80, ::Val{capacity} = Val{256}())

Produce an output (Unsigned or NTuple{n, <:Unsigned}) value, by performing TurboSHAKE on message with a certain capacity and delimsuffix.

See also: ingest, pad, squeeze.

Sponge & Vine

KangarooTwelve.Sponge — Type

Sponge{rate}

A Keccak state that keeps track of the last lane updated.

See also: ingest, pad, and squeeze.

KangarooTwelve.ByteSponge — Type

ByteSponge{rate}

A Keccak state that keeps track of the last byte updated.

See also: ingest, pad, and squeeze.

KangarooTwelve.GerminatingVine — Type

GerminatingVine{rate}

A vine that hasn't started to grow leaves yet (see the k12 docstring).

See also: Vine, ingest, ingest_length, finalise.

KangarooTwelve.Vine — Type

Vine{rate}

A vine with leaves (see the k12 docstring).

See also: GerminatingVine, ingest, ingest_length, finalise.

KangarooTwelve.ingest_length — Function

ingest_length(accumulator::Union{<:ByteSponge, <:AbstractVine}, val::UInt, ::Val{bufsize}=Val{8}())

Ingest a right-encoded form of val into accumulator, allowing the encoding to be up to bufsize bytes.

KangarooTwelve.finalise — Function

finalise(vine::AbstractVine) -> ByteSponge

Finalise vine by folding in the current leaf returning the padded trunk.

State operations

The ingest, pad, squeeze, and squeeze! are implemented for state tuples, sponges, and vines.

KangarooTwelve.ingest — Function

ingest(state::NTuple{25, UInt64}, block::NTuple{rate, UInt64})

Ingest a single block of input (with implied rate) into state or sponge.

The first rate elements of the state are xor'd block, and then the state is permuted with keccak_p1600.

ingest(state::NTuple{25, UInt64}, ::Val{capacity}, message::AbstractVector{<:Unsigned})

Ingest message into state, with the rate calculated based on capacity.

This breaks message into rate-sized blocks and then ingests them (as per ingest(state, ::NTuple{rate, UInt64})) in turn.

ingest(sponge::AbstractSponge, as::Type{<:Unsigned}, leaf::AbstractVector{<:Unsigned})

Ingest leaf into sponge by transforming it to an as via turboshake, and ingesting that result.

ingest(sponge::AbstractSponge, block::AbstractVector{<:Unsigned})

Ingest each element of block into sponge.

ingest(sponge::ByteSponge, x::Union{<:Unsigned, NTuple{N, <:Unsigned}})

Ingest x into sponge.

ingest(vine::AbstractVine, leaflet::AbstractVector{<:Unsigned})
ingest(vine::AbstractVine, x::Unsigned)

Ingest leaflet/x into vine, this may go into the leaves or the trunk depending on the vine type, and may cause the current leaf to be folded into the trunk, and a new leaf grown.

KangarooTwelve.pad — Function

pad(state::NTuple{25, UInt64}, ::Val{capacity}, lastbyte::UInt, delimsuffix::UInt8)

Perform "padding" of the lastbyte byte of state with delimsuffix.

This is the final modification to state in turboshake.

KangarooTwelve.squeeze — Function

squeeze(outtype::Type, state::NTuple{25, UInt64}, ::Val{capacity})

Squeeze an outtype out of state. outtype can be an Unsigned type or an unsigned NTuple.

squeeze(T::Type, sponge::AbstractSponge)

Squeeze a T out of sponge.

KangarooTwelve.squeeze! — Function

squeeze!(output::Vector{UInt64}, state::NTuple{25, UInt64}, ::Val{capacity})

Squeeze state into output.

squeeze!(output::Vector{UInt64}, sponge::AbstractSponge)

Squeeze sponge into output.